[Bash] Rename cat() to ccat()
Overriding cat was a bad idea. Lots of programs call cat and expect the
coreutils executable. Invoking pygmentize instead was breaking a few
things, most notably:
1. Running $ cat file1 file2 > file_cat
2. make <tab>
In order to get back to overriding cat, a couple of things will have to
be taken care of. If we can ensure a way of having the function run ONLY
from the user shell and not when some program invokes cat, it could be
made to work.
Issue 1 can be handled by making it ignore pygmentize when there are
options passed to cat or of there are multiple files.