[Bash] Update Bash Functions and Aliases
Bash Aliases:
* cleantex: Also remove fdb_latexmk and fls files
* pfind: New alias. pgrep only lists all the pids of the matching
processes. But when you don't know which one to kill, use pfind to get
the complete paths too

Bash Functions:
* pacup(): Update function. Use -c in the Wget invokation
* ssh-genkey(): New function. Wrapper to easily generate a new SSH key.
This function takes the name of the server for which a new key is being
made. It creates the relevant directory in .ssh/ and generates a key.
Later it also automatically adds a snippet to the ssh config file so
that the key is used when connecting to that server.