Add a mechanism to create an install db
An install db is a plain text file that contains a list of packages that
have been installed on this system and a mapping of which files are
located where. This is can be used in the future for an uninstall
operation too.

In the future, the dotfiles script will parse this installdb and check
if the current package has already been installed.